Feat/klimatplaner at start (#214)
Add climate plans to start and municipality pages, as well as services and types.
Also general refactoring of a few components, copy and style updates.
