+++ components/FactSection.tsx
@@ -5,12 +5,27 @@ import { H3, Paragraph, ParagraphBold } from './Typography'
 import Icon from '../public/icons/info.svg'
 import IconGreen from '../public/icons/info-green.svg'
 
-
 const Row = styled.div`
   display: flex;
   flex-direction: row;
   justify-content: space-between;
   margin: 0.8rem 0;
+`
+
+const SectionLeft = styled.section`
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+  width: 90%;
+`
+
+const SectionRight = styled.section`
+  text-align: right; 
+`
+
+const InfoHeading = styled(H3)`
+  font-weight: 200;
+  font-size: inherit;
 `
 
 const InfoSection = styled.div`
@@ -26,11 +41,6 @@ const InfoSection = styled.div`
   }
 `
 
-const InfoHeading = styled(H3)`
-  font-weight: 200;
-  font-size: inherit;
-`
-
 const InfoParagraph = styled(Paragraph)`
   margin: 0;
 `
@@ -41,17 +51,6 @@ const StyledIcon = styled.div`
   &:hover {
     cursor: pointer;
   }
-`
-
-const SectionLeft = styled.section`
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-  width: 90%;
-`
-
-const SectionRight = styled.section`
-  text-align: right; 
 `
 
 type Props = {
