+++ data/political/political_rule.py
@@ -4,7 +4,8 @@ import json
 import numpy as np
 import pandas as pd
 
-path_political_rule = './datasets/politicalRule2022.xlsx'
+path_political_rule = 'politicalRule2022.xlsx'
+
 df = pd.read_excel(path_political_rule)
 
 df['Kommun'] = df['Unnamed: 1']
